refactor(auth): split core/auth/__init__.py into submodules
Move logic out of the package __init__ into focused modules:
- keys.py: RSA keypair load/generate
- tokens.py: JWT access/refresh tokens
- cookies.py: zone_auth/zone_refresh cookie helpers
- identity.py: append request-based user/app resolution (validate API
  token, resolve app from token, get_current_user_from_request)

__init__.py becomes a thin re-export shim so existing
`from compute_space.core import auth` callers keep working.
